+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 ### Added
-- Add option to list current devices
+- Add option to list input and block devices
 - Add option to describe device properties
 - Allow to match device based on property set
+### Removed
+- Restriction on subsystem type
 
 ## [0.2.0] - 2018-09-03
 ### Fixed